Backend/requirements.txt
fastapi
pytesseract
Pillow
uvicorn
python-multipart
